backend final

+++ backend/controllers/contactController.js
@@ -57,6 +57,21 @@ controller.updateContact = asyncHandler(async (req ,res) => {
     throw new Error('User not found');
   }
 
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // console.log(contact.user, typeof contact.user);
+  // console.log(req.user.id, typeof req.user.id);
+
+  // Make sure the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true})
   
   res.status(200).json(updatedContact);
@@ -74,6 +89,18 @@ controller.deleteContact = asyncHandler(async (req ,res) => {
   if (!contact) {
     res.status(400);
     throw new Error('User not found');
+  }
+
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   await Contact.findByIdAndRemove(req.params.id)
